Correccion tipo de datos
Se realizo una corrección en los tipos de datos del mapeo de las tablas GTW_COMISION y GTW_COMISION_SEGMENTO

+++ PROYECTO/PaymentGateway/gateway/src/main/java/ec/edu/espe/gateway/comision/controller/ComisionSegmentoController.java
@@ -33,7 +33,7 @@ public class ComisionSegmentoController {
     @GetMapping("/{comision}/{transaccionesDesde}")
     public ResponseEntity<ComisionSegmento> getById(
             @PathVariable Integer comision,
-            @PathVariable BigDecimal transaccionesDesde) {
+            @PathVariable Integer transaccionesDesde) {
         return segmentoService.findById(comision, transaccionesDesde)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -58,8 +58,8 @@ public class ComisionSegmentoController {
     @PutMapping("/{comision}/{transaccionesDesde}")
     public ResponseEntity<ComisionSegmento> update(
             @PathVariable Integer comision,
-            @PathVariable BigDecimal transaccionesDesde,
-            @RequestParam BigDecimal transaccionesHasta,
+            @PathVariable Integer transaccionesDesde,
+            @RequestParam Integer transaccionesHasta,
             @RequestParam BigDecimal monto) {
         try {
             ComisionSegmento updatedSegmento = segmentoService.update(
@@ -79,7 +79,7 @@ public class ComisionSegmentoController {
     @DeleteMapping("/{comision}/{transaccionesDesde}")
     public ResponseEntity<Void> delete(
             @PathVariable Integer comision,
-            @PathVariable BigDecimal transaccionesDesde) {
+            @PathVariable Integer transaccionesDesde) {
         try {
             segmentoService.deleteById(comision, transaccionesDesde);
             return ResponseEntity.noContent().build();
